moved utils function into file from metadata
